separated enemy logic from main and added the ability to send the scene and image to enemy
