+++ EnemyLogic.py
@@ -1,0 +1,22 @@
+import pygame
+
+class Enemy:
+
+    def __init__(self, x, y, hp, range, dmg, cooldown, screen, image):
+        self.hp = hp
+        self.range = range
+        self.dmg = dmg
+        self.cooldown = cooldown
+        self.x = x
+        self.y = y
+        self.image = image
+        self.rect = self.image.get_rect(topleft=(self.x, self.y))
+        self.screen = screen
+
+    def draw(self):
+
+        #pygame.draw.rect(screen, (255, 0, 0), (self.x, self.y, 40, 40))  # Red square as an Enemy
+        self.screen.blit(self.image, (self.x, self.y))
+        pygame.draw.rect(self.screen, (255, 255, 255), (self.x, self.y + 45, 40, 5))  # HP Bar Background White
+        hpPercentage = self.hp / 50 # Max HP: 50 = 100%
+        pygame.draw.rect(self.screen, (0, 255, 0), (self.x, self.y + 45, 40 * hpPercentage, 5))  # HP Bar Green
